split Fixnum and RblFloat into seperate class file
