+++ roscala/src/main/scala/coop/rchain/rosette/VirtualMachine.scala
@@ -2,7 +2,7 @@ package coop.rchain.rosette
 
 import com.typesafe.scalalogging.Logger
 import coop.rchain.rosette.Ob._
-import coop.rchain.rosette.prim.Prim
+import coop.rchain.rosette.prim.{Fixnum, Prim}
 
 sealed trait Work
 case object NoWorkLeft extends Work
