+++ roscala/src/main/scala/coop/rchain/rosette/prim/package.scala
@@ -2,11 +2,11 @@ package coop.rchain.rosette
 
 package object prim {
   // For now keys in Prims correspond to Prim::primnum in Rosette
-  val Prims = Map[Int, Prim](197 -> Number.flPlus,
-                             222 -> Number.fxMod,
-                             223 -> Number.fxDiv,
-                             224 -> Number.fxTimes,
-                             225 -> Number.fxMinus,
-                             226 -> Number.fxPlus)
+  val Prims = Map[Int, Prim](197 -> RblFloat.flPlus,
+                             222 -> Fixnum.fxMod,
+                             223 -> Fixnum.fxDiv,
+                             224 -> Fixnum.fxTimes,
+                             225 -> Fixnum.fxMinus,
+                             226 -> Fixnum.fxPlus)
 
 }
